Add files via upload
There are 3 departments within the system network:
●	Research and Development;
●	Information Technology;
●	HR and Personnel.

Static IP – the PC have different IP and send request to Router and Router changes the IP Address to be more secured, after the Router sends the packages to Server, then from Server the answer is coming to Router because he don't know the PC that is asking for the answer, so Router knows the address and the package is going from Router to our PC address.
RIR – Regional Internet Registry for all public Internet Addresses must be registered:
a.	10.0. - private A class;
b.	172. - private B class;
c.	192. - private C class.

Port address Translation (PAT) – identifying the type of the traffic going in the computer
IF using a port is better to access the internet, because the server knows your request from the Port address. Hiding ourselves to have a port, so the same IP address and using 100 computers assigned to 1 computer server, and those 100 computers have different PAT.
80 - is for http;
25 - is for email;
16 bit - 65535 ports.